added/tablet styles
